% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_prepare.R
\name{c_prepare}
\alias{c_prepare}
\title{Prepare data for C API: list of row-vectors with NA indicators}
\usage{
c_prepare(data)
}
\arguments{
\item{data}{A dataframe to be prepared for the C API.}
}
\value{
A list of row-vectors, where each vector contains 0s (non-NA) and 1s (NA).
}
\description{
This function prepares the data to be used by a C API by transforming each row of the dataframe into a vector
of 0s and 1s, where 0 indicates a non-NA value and 1 indicates an NA value.
}
\keyword{internal}
