% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_prepare.R
\name{check_input}
\alias{check_input}
\title{Check that the input is a valid dataframe}
\usage{
check_input(data)
}
\arguments{
\item{data}{The input to be checked.}
}
\value{
NULL. The function stops with an error message if the input is not a valid dataframe.
}
\description{
This function checks if the input is a valid dataframe with at least two rows and two columns.
It throws an error if the input does not meet these conditions.
}
\keyword{internal}
