% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_prepare.R
\name{drop_prepare}
\alias{drop_prepare}
\title{internal function to detect dropouts}
\usage{
drop_prepare(data, last_col)
}
\arguments{
\item{data}{dataframe or tibble to detect dropouts.}

\item{last_col}{Index position or column name of the last survey item. This is an optional argument and is needed if there are other columns in your dataframe after the survey items you want to detect the dropout for.}
}
\value{
A list containing a dataframe or tibble with the following for each row of your dataset: A logical whether a dropout has been detected and, if so, the column name where it occurs and the index of the column. As well as the value of the col_index
}
\description{
internal function to detect dropouts
}
