% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropout-package.R
\docType{package}
\name{dropout-package}
\alias{dropout}
\alias{dropout-package}
\title{dropout: Handling Incomplete Responses in Survey Data Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Offers robust tools to identify and manage incomplete responses in survey datasets, thereby enhancing the quality and reliability of research findings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hendr1km/dropout}
  \item \url{https://hendr1km.github.io/dropout/}
  \item Report bugs at \url{https://github.com/hendr1km/dropout/issues}
}

}
\author{
\strong{Maintainer}: Hendrik Mann \email{hendrik.mann@uni-wuppertal.de}

}
\keyword{internal}
