% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_column}
\alias{metric_column}
\title{Compute column name for dropped observations}
\usage{
metric_column(c_output, drop_index, data)
}
\arguments{
\item{c_output}{A list containing output data to detect dropped columns.}

\item{drop_index}{A vector indicating the index of the dropped column for each row.}

\item{data}{The original dataset.}
}
\value{
A vector of column names corresponding to dropped values.
}
\description{
This function returns the column names corresponding to the dropped values in the dataset.
}
\keyword{internal}
