% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_complete}
\alias{metric_complete}
\title{Compute completeness of each column}
\usage{
metric_complete(na, data)
}
\arguments{
\item{na}{A vector of NA counts for each column.}

\item{data}{The original dataset.}
}
\value{
A vector of completeness values (proportions) for each column.
}
\description{
This function calculates the completeness of each column in the dataset by subtracting the NA count from the total number of rows.
}
\keyword{internal}
