% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_drop}
\alias{metric_drop}
\title{Compute drop metric for each column}
\usage{
metric_drop(c_output)
}
\arguments{
\item{c_output}{A list containing output data to compute the drop metric.}
}
\value{
A vector of drop values.
}
\description{
This function computes the number of drops for each column based on the \code{c_output} data.
}
\keyword{internal}
