% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_drop_id}
\alias{metric_drop_id}
\title{Compute drop indicator for each row}
\usage{
metric_drop_id(c_output)
}
\arguments{
\item{c_output}{A list containing output data to compute the drop indicator.}
}
\value{
A logical vector indicating dropped rows (TRUE = dropped).
}
\description{
This function computes a logical vector indicating whether each row in the dataset has been dropped.
}
\keyword{internal}
