% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_drop_index}
\alias{metric_drop_index}
\title{Compute drop index for each row}
\usage{
metric_drop_index(c_output)
}
\arguments{
\item{c_output}{A list containing output data to compute the drop index.}
}
\value{
A vector of drop indices (or NA if no drop occurred).
}
\description{
This function returns the index of the dropped column for each row, or \code{NA} if no column was dropped.
}
\keyword{internal}
