% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_na}
\alias{metric_na}
\title{Compute total NA count for each column}
\usage{
metric_na(drop, sec_na, single_na)
}
\arguments{
\item{drop}{A vector of drop counts for each column.}

\item{sec_na}{A vector of section NA counts for each column.}

\item{single_na}{A vector of single NA counts for each column.}
}
\value{
A vector of total NA counts.
}
\description{
This function calculates the total NA count for each column by summing the drop, section NA, and single NA metrics.
}
\keyword{internal}
