% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_sec_length}
\alias{metric_sec_length}
\title{Compute section length metric for each column}
\usage{
metric_sec_length(c_output)
}
\arguments{
\item{c_output}{A list containing output data to compute the section length metric.}
}
\value{
A vector of section length values.
}
\description{
This function calculates the mode (most common value) of the left-out section length for each column.
}
\keyword{internal}
