% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_sec_na}
\alias{metric_sec_na}
\title{Compute section NA metric for each column}
\usage{
metric_sec_na(c_output)
}
\arguments{
\item{c_output}{A list containing output data to compute the section NA metric.}
}
\value{
A vector of section NA values.
}
\description{
This function calculates the total number of section NAs for each column in the dataset.
}
\keyword{internal}
