% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_single_na}
\alias{metric_single_na}
\title{Compute single NA metric for each column}
\usage{
metric_single_na(c_output)
}
\arguments{
\item{c_output}{A list containing output data to compute the single NA metric.}
}
\value{
A vector of single NA values.
}
\description{
This function calculates the total number of single NAs for each column.
}
\keyword{internal}
