% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics_detect}
\alias{metrics_detect}
\title{Detect dropped observations and columns}
\usage{
metrics_detect(data, c_output)
}
\arguments{
\item{data}{The original dataset.}

\item{c_output}{A list containing output data to detect dropped values.}
}
\value{
A dataframe indicating the dropped rows and columns.
}
\description{
This function identifies rows and columns with dropped values in the dataset based on the \code{c_output}.
}
\keyword{internal}
