% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics_summary}
\alias{metrics_summary}
\title{Generate a summary of metrics for the dataset}
\usage{
metrics_summary(data, c_output)
}
\arguments{
\item{data}{A dataframe containing the dataset.}

\item{c_output}{A list or dataframe containing the computed output that helps generate the metrics.}
}
\value{
A dataframe summarizing the computed metrics for each column.
}
\description{
This function calculates a variety of metrics on the provided data and a corresponding output (\code{c_output}).
These metrics include drop rate, section NA, section length, single NA, and the number of complete rows.
}
\keyword{internal}
